move python stuff out of the way
